Use `StaticResourceLoader` in the grace-bootstrap
