+++ grace-plugin-api/src/main/groovy/org/grails/plugins/BinaryGrailsPlugin.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2023 the original author or authors.
+ * Copyright 2011-2025 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -40,8 +40,8 @@ import grails.io.IOUtils;
 import grails.plugins.exceptions.PluginException;
 import grails.util.BuildSettings;
 
-import org.grails.core.io.StaticResourceLoader;
 import org.grails.io.support.GrailsResourceUtils;
+import org.grails.io.support.StaticResourceLoader;
 
 /**
  * Models a pre-compiled binary plugin.
@@ -49,6 +49,7 @@ import org.grails.io.support.GrailsResourceUtils;
  * @see grails.plugins.GrailsPlugin
  *
  * @author Graeme Rocher
+ * @author Michael Yan
  * @since 2.0
  */
 public class BinaryGrailsPlugin extends DefaultGrailsPlugin {
@@ -237,8 +238,7 @@ public class BinaryGrailsPlugin extends DefaultGrailsPlugin {
         Resource url = this.baseResourcesResource;
         Properties properties = null;
         if (url != null) {
-            StaticResourceLoader resourceLoader = new StaticResourceLoader();
-            resourceLoader.setBaseResource(url);
+            StaticResourceLoader resourceLoader = new StaticResourceLoader(url);
             ResourcePatternResolver resolver = new PathMatchingResourcePatternResolver(resourceLoader);
             try {
                 // first load all properties
